북마크 없을시 뷰 수정
