+++ core/ui/src/main/java/com/example/sachosaeng/core/ui/IntConstant.kt
@@ -1,0 +1,5 @@
+package com.example.sachosaeng.core.ui
+
+object IntConstant {
+    const val BOTTOM_BAR_HEIGHT = 76
+}
