+++ feature/bookmark/src/main/java/com/example/sachosaeng/feature/bookmark/BookmarkScreen.kt
@@ -10,6 +10,7 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyRow
@@ -42,6 +43,9 @@ import com.example.sachosaeng.core.ui.theme.Gs_G5
 import com.example.sachosaeng.core.ui.theme.Gs_White
 import com.example.sachosaeng.core.util.constant.IntConstant.ALL_CATEGORY_ID
 import com.example.sachosaeng.core.util.extension.StringExtension.toColorResource
+import com.example.sachosaeng.feature.bookmark.component.BookmarkEmptyScreen
+import com.example.sachosaeng.feature.bookmark.component.BookmarkList
+import com.example.sachosaeng.feature.bookmark.component.CategoryRow
 import org.orbitmvi.orbit.compose.collectAsState
 
 @Composable
@@ -117,16 +121,10 @@ internal fun BookmarkScreen(
                     )
                 },
                 {
-                    CategoryRow(
-                        categories = state.allCategory,
+                    ArticleTabScreen(
+                        state = state,
                         onModifyButtonClicked = onModifyButtonClicked,
-                        isModifyMode = state.isModifyMode,
-                    )
-                    BookmarkList(
-                        isModifyMode = state.isModifyMode,
-                        bookmarks = state.bookmarkedArticleList,
-                        onBookmarkClicked = onBookmarkedArticleClick,
-                        selectedForModifyBookmarkList = state.selectedForModifyBookmarkList,
+                        onBookmarkedArticleClick = onBookmarkedArticleClick,
                         onSelectForModifyBookmark = onSelectForModifyBookmark
                     )
                 },
@@ -152,6 +150,9 @@ private fun VoteTabScreen(
             onCategoryClicked = onCategoryClicked,
             isModifyMode = state.isModifyMode
         )
+        if (state.bookmarkedVoteList.isEmpty()) {
+            BookmarkEmptyScreen(emptyLabel = stringResource(id = R.string.bookmarked_vote_is_empty_description))
+        }
         BookmarkList(
             isModifyMode = state.isModifyMode,
             bookmarks = state.bookmarkedVoteList,
@@ -170,89 +171,27 @@ private fun VoteTabScreen(
 }
 
 @Composable
-fun CategoryRow(
-    selectedCategory: Category? = null,
-    categories: List<Category>,
-    modifier: Modifier = Modifier,
-    onCategoryClicked: (Category) -> Unit = {},
+fun ArticleTabScreen(
+    state: BookmarkScreenUiState,
     onModifyButtonClicked: () -> Unit = {},
-    isModifyMode: Boolean = false
+    onBookmarkedArticleClick: (Bookmark) -> Unit = {},
+    onSelectForModifyBookmark: (Bookmark) -> Unit = {}
 ) {
-    Box {
-        LazyRow(
-            modifier = modifier.padding(top = 16.dp, bottom = 16.dp, end = 60.dp),
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            items(categories.size) {
-                CategoryCard(
-                    isSelected = selectedCategory == categories[it],
-                    category = categories[it],
-                    onCategoryClicked = onCategoryClicked
-                )
-            }
-        }
-        Row(
-            modifier = Modifier
-                .background(Gs_G2)
-                .padding(16.dp)
-                .align(Alignment.CenterEnd)
-                .noRippleClickable { onModifyButtonClicked() }
-        ) {
-            Text(
-                color = if (isModifyMode) Gs_G5 else Gs_Black,
-                text = stringResource(id = if (isModifyMode) R.string.cancel_label else R.string.modify_label),
-                fontSize = 14.sp,
-                fontWeight = FontWeight.W500
-            )
-        }
+    CategoryRow(
+        categories = state.allCategory,
+        onModifyButtonClicked = onModifyButtonClicked,
+        isModifyMode = state.isModifyMode,
+    )
+    if (state.bookmarkedArticleList.isEmpty()) {
+        BookmarkEmptyScreen(emptyLabel = stringResource(id = R.string.bookmarked_article_is_empty_description))
     }
-}
-
-@Composable
-private fun CategoryCard(
-    modifier: Modifier = Modifier,
-    isSelected: Boolean = false,
-    category: Category,
-    onCategoryClicked: (Category) -> Unit = {}
-) {
-    Card(
-        colors = CardDefaults.cardColors().copy(
-            containerColor = Color(category.color.toColorResource()),
-            contentColor = Color(category.textColor.toColorResource())
-        ),
-        shape = RoundedCornerShape(4.dp),
-        modifier = modifier
-            .border(
-                1.dp,
-                if (isSelected) Color(category.textColor.toColorResource()) else Color.Transparent,
-                RoundedCornerShape(4.dp)
-            )
-            .clickable {
-                onCategoryClicked(category)
-            },
-    ) {
-        Row(
-            modifier = modifier
-                .padding(10.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            if (category.imageUrl?.isNotEmpty() == true) {
-                AsyncImage(
-                    alignment = Alignment.CenterEnd,
-                    contentDescription = "", model = category.imageUrl,
-                    modifier = Modifier
-                        .size(18.dp)
-                        .padding(end = 8.dp)
-                )
-            }
-            Text(
-                text = category.name,
-                fontSize = 15.sp,
-                fontWeight = FontWeight.W600
-            )
-        }
-    }
+    BookmarkList(
+        isModifyMode = state.isModifyMode,
+        bookmarks = state.bookmarkedArticleList,
+        onBookmarkClicked = onBookmarkedArticleClick,
+        selectedForModifyBookmarkList = state.selectedForModifyBookmarkList,
+        onSelectForModifyBookmark = onSelectForModifyBookmark
+    )
 }
 
 @Preview
@@ -284,26 +223,7 @@ fun BookmarkScreenPreview() {
                     textColor = "#FFFFFF"
                 ),
             ),
-            bookmarkedVoteList = listOf(
-                Bookmark(
-                    bookmarkId = 1,
-                    id = 1,
-                    title = "Bookmark1",
-                    description = "Description1"
-                ),
-                Bookmark(
-                    bookmarkId = 2,
-                    id = 2,
-                    title = "Bookmark2",
-                    description = "Description2"
-                ),
-                Bookmark(
-                    bookmarkId = 3,
-                    id = 3,
-                    title = "Bookmark3",
-                    description = "Description3"
-                ),
-            ),
+            bookmarkedVoteList = listOf(),
             isModifyMode = true,
             selectedForModifyBookmarkList = listOf(
                 Bookmark(
